Release version 3.5.0

+++ src/main/java/uk/ac/soton/itinnovation/security/modelquerier/EntityCache.java
@@ -186,13 +186,20 @@ public class EntityCache {
      * triple store if the cache is synchronised.
      */
     public <T extends EntityDB> void storeEntity(T entity, String typeKey, String graph) {
+        // Get the entity URI
+        String uri = entity.getUri();
+
         // Add the entity to the cache without changing validity status for the entity type
         cacheEntity(entity, typeKey, false, graph);
 
         // Add it to the staging area, so we know it must be written to the DB at the next synchronisation 
         Map<String, Map<String, EntityDB>> storeByType = storeByTypeByGraph.computeIfAbsent(graph, k -> new HashMap<>());
         Map<String, EntityDB> store = storeByType.computeIfAbsent(typeKey, k -> new HashMap<>());
-        store.put(entity.getUri(), entity);
+        store.put(uri, entity);
+
+        // If the same entity is in the list of entities to be deleted at the next synchronisation, remove it
+        Map<String, EntityDB> deletions = deleteByType.computeIfAbsent(typeKey, k -> new HashMap<>());
+        if(deletions.keySet().contains(uri)) deletions.remove(uri);
 
     }
 
@@ -228,7 +235,7 @@ public class EntityCache {
      * Deletion is only permitted on system model entities, and must be done across all the
      * system model graphs.
      */
-    public <T extends EntityDB> void deleteEntity(T entity, String typeKey) {
+    public <T extends EntityDB> void deleteEntity(T entity, String typeKey, Boolean skipDelete) {
         // Get the URI
         String uri = entity.getUri();
         
@@ -250,12 +257,18 @@ public class EntityCache {
             }
         }
 
-        // Add it to the staging area for entities to be deleted, so we know to delete it 
-        Map<String, EntityDB> deletions = deleteByType.computeIfAbsent(typeKey, k -> new HashMap<>());
-        deletions.put(entity.getUri(), entity);
+        if(!skipDelete){
+            // Add it to the staging area for entities to be deleted, so we know to delete it 
+            Map<String, EntityDB> deletions = deleteByType.computeIfAbsent(typeKey, k -> new HashMap<>());
+            deletions.put(entity.getUri(), entity);    
+        }
 
         // Do not remove it from the list of valid cached entities - deleted means a null response is valid
 
+    }
+
+    public <T extends EntityDB> void deleteEntity(T entity, String typeKey){
+        deleteEntity(entity, typeKey, false);
     }
 
     /* Gets all entities of a given type from the cache for a specified list of graphs.
@@ -351,20 +364,7 @@ public class EntityCache {
         return valid;
     }
 
-    /* Clears entities from the delete list if they were subsequently stored */
-    public void prepareSync(){
-        for(String graph : systemGraphs) {
-            Map<String, Map<String, EntityDB>> cacheByType = cacheByTypeByGraph.computeIfAbsent(graph, k -> new HashMap<>());
-            for(String typeKey : cacheByType.keySet()){
-                Map<String, EntityDB> deletions = deleteByType.computeIfAbsent(typeKey, k -> new HashMap<>());
-                Map<String, EntityDB> cacheThisType = cacheByType.computeIfAbsent(typeKey, k -> new HashMap<>());
-                for(String uri : cacheThisType.keySet()){
-                    if(deletions.keySet().contains(uri)) deletions.remove(uri);
-                }
-            }
-        }
-    }
-
+    /* Clears all content from the cache */
     public void clear() {
         cacheByTypeByGraph.clear();
         cacheByUriByGraph.clear();
